perf: 调整辅助工具及核心逻辑(#0)

+++ packages/mini-helper/src/index.ts
@@ -2,9 +2,11 @@ import path from "path";
 import process from "process";
 import fs from "fs";
 import chalk from "chalk";
+import { MiniCIDefaultDir, Platform } from "@hengshuai/mini-type";
 
 export const rootDir = process.cwd();
 
+/** 取ci执行目录 */
 export const resolvePath = (p: string) => path.resolve(rootDir, p);
 
 export function sleep(ms = 500) {
@@ -14,12 +16,20 @@ export function sleep(ms = 500) {
 // 删除文件
 export async function deleteFile(filePath: string) {
   try {
-    if (filePath && filePath !== "") {
+    if (filePath && filePath !== "" && fs.existsSync(resolvePath(filePath))) {
       await fs.unlinkSync(resolvePath(filePath));
     }
   } catch (err) {
     console.error(err);
   }
+}
+
+/**
+ * 创建图片文件名字
+ * @param platform 平台
+ */
+export function createImgFileName(platform?: Platform) {
+  return `${MiniCIDefaultDir}/${platform ? platform : "default"}__${+new Date()}.png`;
 }
 
 /**
